feat : #29 Tap의 상태를 전역이 아닌 훅을 만들어서 관리
