+++ src/stores/recommend.store.ts
@@ -4,10 +4,8 @@ import { persist } from "zustand/middleware";
 type RecommendStore = {
   cityId: number;
   countryId: number;
-  currentTab: string;
   setCityId: (cityId: number) => void;
   setCountryId: (countryId: number) => void;
-  setCurrentTab: (currentTab: string) => void;
 };
 
 const useRecommendStore = create<RecommendStore>()(
@@ -15,10 +13,8 @@ const useRecommendStore = create<RecommendStore>()(
     (set) => ({
       cityId: 1,
       countryId: 1,
-      currentTab: "acommodation",
       setCityId: (cityId: number) => set({ cityId }),
       setCountryId: (countryId: number) => set({ countryId }),
-      setCurrentTab: (currentTab: string) => set({ currentTab }),
     }),
     {
       name: "recommend",
